tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Integration tests for the Exasol launcher tool"
authors = [
    {name = "Exasol", email = "entwicklung@exasol.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "semver (>=3.0.4,<4.0.0)",
    "requests (>=2.31.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "tests"}]


[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
ruff = "^0.12.9"
pydantic = {extras = ["mypy"], version = "^2.11.7"}
types-requests = "^2.31.0"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(funcName)s: %(message)s"

[tool.mypy]
python_version = "3.13"
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_optional = true
no_implicit_reexport = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_any_decorated = true
disallow_subclassing_any = true

plugins = [
    "pydantic.mypy",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # These force us to document every
    # single function and module.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",

    # Conflicting rules
    "D203", # `incorrect-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
    "D213", # `multi-line-summary-first-line` (D212) and `multi-line-summary-second-line` (D213) are incompatible.

    "S603", # Frequent false positives
    "S101", # Forbids asserts that we need
    "LOG015", # Forbids usage of the global logger

    # Conflicts with formatter
    "COM812", # The following rules may cause conflicts when used with the formatter: `COM812`
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
